CombinedOrch/requirements.txt
openai==1.3.5
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
numpy==1.24.3
scikit-learn==1.3.2
requests==2.31.0
python-multipart==0.0.6
aiofiles==23.2.1
mem0ai==0.1.114 